refactor(plugins): add file extension on import paths
