+++ src/plugins/AutoHostSelector.ts
@@ -1,8 +1,10 @@
-import { escapeUserName, Lobby } from '..';
-import { BanchoResponseType, MpSettingsResult } from '../parsers';
-import { Player, revealUserName, disguiseUserName } from '../Player';
-import { Disposable, TypedEvent } from '../libs';
-import { LobbyPlugin } from './LobbyPlugin';
+import { escapeUserName } from '../Player.js';
+import { Lobby } from '../Lobby.js';
+import { BanchoResponseType,  } from '../parsers/CommandParser.js';
+import { MpSettingsResult } from '../parsers/MpSettingsParser.js';
+import { Player, revealUserName, disguiseUserName } from '../Player.js';
+import { Disposable, TypedEvent } from '../libs/TypedEvent.js';
+import { LobbyPlugin } from './LobbyPlugin.js';
 import config from 'config';
 import log4js from 'log4js';
