+++ src/plugins/MatchAborter.ts
@@ -1,9 +1,10 @@
-import { Lobby } from '..';
-import { MpSettingsResult, StatStatuses } from '../parsers';
-import { Player, MpStatuses } from '../Player';
-import { LobbyPlugin } from './LobbyPlugin';
+import { Lobby } from '../Lobby.js';
+import { MpSettingsResult } from '../parsers/MpSettingsParser.js';
+import { StatStatuses } from '../parsers/StatParser.js';
+import { Player, MpStatuses } from '../Player.js';
+import { LobbyPlugin } from './LobbyPlugin.js';
 import config from 'config';
-import { VoteCounter } from './VoteCounter';
+import { VoteCounter } from './VoteCounter.js';
 
 export interface MatchAborterOption {
   vote_rate: number; // アボート投票時の必要数/プレイヤー数
